+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -13,15 +13,19 @@ import mozilla.components.browser.engine.gecko.webextension.GeckoWebExtension
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy
+import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy.TrackingCategory
 import mozilla.components.concept.engine.EngineSession.SafeBrowsingPolicy
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.engine.Settings
+import mozilla.components.concept.engine.content.blocking.TrackerLog
 import mozilla.components.concept.engine.history.HistoryTrackingDelegate
 import mozilla.components.concept.engine.mediaquery.PreferredColorScheme
 import mozilla.components.concept.engine.utils.EngineVersion
 import mozilla.components.concept.engine.webextension.WebExtension
 import org.json.JSONObject
+import org.mozilla.geckoview.ContentBlockingController
+import org.mozilla.geckoview.ContentBlockingController.Event
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoRuntime
 import org.mozilla.geckoview.GeckoRuntimeSettings
@@ -52,6 +56,37 @@ class GeckoEngine(
             @Suppress("TooGenericExceptionThrown")
             throw RuntimeException("GeckoRuntime is shutting down")
         }
+    }
+
+    /**
+     * Fetch a list of trackers logged for a given [session] .
+     *
+     * @param session the session where the trackers were logged.
+     * @param onSuccess callback invoked if the data was fetched successfully.
+     * @param onError (optional) callback invoked if fetching the data caused an exception.
+     */
+    override fun getTrackersLog(
+        session: EngineSession,
+        onSuccess: (List<TrackerLog>) -> Unit,
+        onError: (Throwable) -> Unit
+    ) {
+        val geckoSession = (session as GeckoEngineSession).geckoSession
+        runtime.contentBlockingController.getLog(geckoSession).then({ contentLogList ->
+            val list = contentLogList ?: emptyList()
+            val logs = list.map { logEntry ->
+                logEntry.toTrackerLog()
+            }.filterNot {
+                !it.cookiesHasBeenBlocked &&
+                    it.blockedCategories.isEmpty() &&
+                    it.loadedCategories.isEmpty()
+            }
+
+            onSuccess(logs)
+            GeckoResult<Void>()
+        }, { throwable ->
+            onError(throwable)
+            GeckoResult<Void>()
+        })
     }
 
     /**
@@ -170,7 +205,7 @@ class GeckoEngine(
                 value?.let { policy ->
                     val activateStrictSocialTracking =
                         policy.strictSocialTrackingProtection ?: policy.trackingCategories.contains(
-                            TrackingProtectionPolicy.TrackingCategory.STRICT
+                            TrackingCategory.STRICT
                         )
 
                     runtime.settings.contentBlocking.setStrictSocialTrackingProtection(
@@ -191,7 +226,7 @@ class GeckoEngine(
              * remove its value from the valid anti tracking categories, when is present.
              */
             val total = trackingCategories.sumBy { it.id }
-            return if (contains(TrackingProtectionPolicy.TrackingCategory.SCRIPTS_AND_SUB_RESOURCES)) {
+            return if (contains(TrackingCategory.SCRIPTS_AND_SUB_RESOURCES)) {
                 total - TrackingProtectionPolicy.TrackingCategory.SCRIPTS_AND_SUB_RESOURCES.id
             } else {
                 total
@@ -276,4 +311,42 @@ class GeckoEngine(
             this.fontSizeFactor = it.fontSizeFactor
         }
     }
+}
+
+internal fun ContentBlockingController.LogEntry.BlockingData.hasBlockedCookies(): Boolean {
+    return category == Event.COOKIES_BLOCKED_BY_PERMISSION ||
+        category == Event.COOKIES_BLOCKED_TRACKER ||
+        category == Event.COOKIES_BLOCKED_ALL ||
+        category == Event.COOKIES_PARTITIONED_FOREIGN ||
+        category == Event.COOKIES_BLOCKED_FOREIGN
+}
+
+internal fun ContentBlockingController.LogEntry.BlockingData.getBlockedCategory(): TrackingCategory {
+    return when (category) {
+        Event.BLOCKED_FINGERPRINTING_CONTENT -> TrackingCategory.FINGERPRINTING
+        Event.BLOCKED_CRYPTOMINING_CONTENT -> TrackingCategory.CRYPTOMINING
+        Event.BLOCKED_SOCIALTRACKING_CONTENT -> TrackingCategory.MOZILLA_SOCIAL
+        Event.BLOCKED_TRACKING_CONTENT -> TrackingCategory.SCRIPTS_AND_SUB_RESOURCES
+        else -> TrackingCategory.NONE
+    }
+}
+
+internal fun ContentBlockingController.LogEntry.BlockingData.getLoadedCategory(): TrackingCategory {
+    return when (category) {
+        Event.LOADED_FINGERPRINTING_CONTENT -> TrackingCategory.FINGERPRINTING
+        Event.LOADED_CRYPTOMINING_CONTENT -> TrackingCategory.CRYPTOMINING
+        Event.LOADED_SOCIALTRACKING_CONTENT -> TrackingCategory.MOZILLA_SOCIAL
+        Event.LOADED_TRACKING_CONTENT -> TrackingCategory.SCRIPTS_AND_SUB_RESOURCES
+        else -> TrackingCategory.NONE
+    }
+}
+
+internal fun ContentBlockingController.LogEntry.toTrackerLog(): TrackerLog {
+    val cookiesHasBeenBlocked = this.blockingData.any { it.hasBlockedCookies() }
+    return TrackerLog(
+        url = origin,
+        loadedCategories = blockingData.map { it.getLoadedCategory() }.filterNot { it == TrackingCategory.NONE },
+        blockedCategories = blockingData.map { it.getBlockedCategory() }.filterNot { it == TrackingCategory.NONE },
+        cookiesHasBeenBlocked = cookiesHasBeenBlocked
+    )
 }